Function Rearranging, GETTIME, SQRT, Connections
Rearranged order of the function definitions in Evaluator to match the
ordering of the list used for detecting function names, which was also
rearranged for more consistency. Should make it easier to find where
things are as more functions are added.

Added simple GETTIME function(unix epoch)

Added SQRT math function

Added REMOVEDATA

Updated Connection class to be consistent with Node implimentation. Also
added leftname and rightname as hard coded values in getConnectionData
in a similar vein to name, for getting the left and right node names of
that connection

Added the rest of the basic connection functions in Evaluator