Parameter List Crosstalk Fix
Commas now completely seperate parameters. When a comma is encountered
it now creates an entirely new list and pushes it onto the parameter
stack, but unlike an opening parenthesis the height is not changed. Then
when a closing parenthesis is found it merges the topmost stack items
with the same height before evaluating a function.

This completely eliminates crosstalk bugs I was having where I could do
something like AVG(2,-2,3) and would result in 1.5 instead of 1, as 2 -
2 would happen and it'd average 0 and 3.

Added some basic error raising