Unique Keys, Decoupled Tokenizer
Added a list keylist that maintains all keys used in the supergraph.
This is used to prevent using the same name for multiple things. This is
done to set the stage for using the pointer operation "." in order to
allow the Evaluator to reference things in the Supergraph.

Added a very basic PRINTKEYS function that simply prints all keys in the
Supergraph. Will probably replace it with GETKEYS or something down the
line

Added in == and != functionality in evaluateFunction that was overlooked

Renamed evaluateExpression to evaluateTokens, then made a new
evaluateExpression that calls a tokenizeExpression then an
evaluateTokens. This was done to decouple tokenizing from evaluating
said tokens, as if you want to run a query over a large number of items
you'd be performing lots of redundant tokenizations of the same
expression.