Data Pointers
Added in basic data pointers similar to other languages. Foo.Bar returns
variable Bar from object Foo, and returns an error if it doesn't exist.
This is very similar to the function GETXDATA except it looks at a
specific object of unspecified type, throws errors, and may eventually
allow for chaining, whereas the function looks at sets of specific
objects and will simply return empty lists if the data doesnt exist.

If a key of an object is given with no pointers, it simply returns that
specific object as its object's return type. For instance nodename would
return nodename of type N, allowing you to use it in something like
REMOVENODES without having to use the string "nodename"

Added verifyKeys which functions the same as its node and connection
based counterparts, except looking accross all the keys instead of a
specific type