Added Queries
Made evaluateTokens actually return something

Added QUERY, which can be used on a group of nodes or connections. Will
only return the objects that have the given expression return as true.
The keyword THIS allows you to reference data within the object being
queried.

Updated PRINT to strip the list structure when given a single item. This
simply improves the look and would make PRINT("test") go from ["test"]
to "test" in when outputted