Made it easier to get connections from nodes
Added dictionary nodeconnectionlist to Supergraph which maintains a list
of connection keys for any given node.

Updated existing methods to take into account this variable so it
properly updates as new connections/nodes are added/removed.

Fixed a bug in removeConnections where removing a recursive connection
would try to remove it twice

Added L support for REMOVENODES and REMOVECONNECTIONS