Static Supergraph, File/Error Handling, Bug Fixes
Changed Supergraph from instantiated to static. I decided there should
only ever be one database at a time for the moment

Added FileHandler. Can do a basic write graph to file and read graph
from file. Can't handle lists currently.

Removed built in error system I was attempting and moved to Python's
exception handling

Created ArtPointsFinder. Is intended to find articulated points given a
graph or a set of nodes and connections. Will implement later

Removed some commented out code from main

Fixed bug where adding a null in SUM would not return an exception

Fixed bug in Supergraph.removeConnections where if a node was removed
before the connection attached to it was it would cause an exception in
the lookup when removing itself from the dictionaries

Tabs are now treated as part of the tokenization process

Supergraph.addConnection and Supergraph.addNode returns true if it
succeeds and false if it fails

Removed Evaulator.tokenizeExpression as it was deprecated by
Evaluator.tokenizeExpressionRegex

Added hardcoded lists in Evaluator for operators and reserved characters

Added Evaluator.checkTokenLegality to see if a token is not a reserved
keyword or contains a reserved character

Removed some redundant parantheses and semi-colons