Updated addNode and addConnection
Supergraph addNode and addConnection will now return the key to the
object they've created, or if it exists, will return the key of the
existing object

They will interact with KeyGenerator to create a new key(an autonumber
appended to the requested key name) if either add_if_exists in the
method call is true or the respective configs are set to true in the
files. These are true by default currently.